outbound/requirements.txt
google-cloud-bigquery
python-dotenv
requests
